[DF] Build ranges with free functions
The `RangesBuilder` class has been removed in favor of having free functions that serve the same purpose.
`get_clustered_ranges` creates ranges for distributed execution based on clusters of some TTree/TChain base dataset. It is called by `TreeHeadNode`
`get_balanced_ranges` creates roughly equal-sized chunks of entries. It is called by `EmptySourceHeadNode`

The `build_range` methods of the head node classes now call directly the free functions in the `Ranges.py` module.
In `TreeHeadNode`, split the work in two steps: first gather the clusters and do some checks against the current value of `npartitions`, then actually compute the ranges.
